More refactoring of serialization types
